requirements.txt
# azure-ai-formrecognizer==3.3.2  # Optional - not needed for core functionality
azure-ai-textanalytics==5.3.0
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
openai==1.108.1
python-dotenv==1.0.0
streamlit==1.28.1
pypdf2==3.0.1
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3
